Fix overflow. Fixes #12

+++ src/cc1101.cc
@@ -99,8 +99,8 @@ Status Radio::setFrequency(double freq) {
 Status Radio::setFrequencyDeviation(double dev) {
   double xosc = CC1101_CRYSTAL_FREQ * 1000;
 
-  int devMin = (xosc / (1 << 17)) * (8 + 0) * 1;
-  int devMax = (xosc / (1 << 17)) * (8 + 7) * (1 << 7);
+  uint32_t devMin = (xosc / ((uint32_t)1 << 17)) * (8 + 0) * 1;
+  uint32_t devMax = (xosc / ((uint32_t)1 << 17)) * (8 + 7) * (1 << 7);
 
   if (dev < devMin || dev > devMax) {
     return STATUS_INVALID_PARAM;
@@ -111,7 +111,7 @@ Status Radio::setFrequencyDeviation(double dev) {
 
   for (uint8_t e = 0; e <= 7; e++) {
     for (uint8_t m = 0; m <= 7; m++) {
-      double t = (xosc / (double)(1ULL << 17)) * (8 + m) * (double)(1ULL << e);
+      double t = (xosc / (double)((uint32_t)1 << 17)) * (8 + m) * (double)((uint32_t)1 << e);
       if (fabs(dev - t) < diff) {
         diff = fabs(dev - t);
         bestE = e;
@@ -133,8 +133,8 @@ void Radio::setChannel(uint8_t ch) {
 Status Radio::setChannelSpacing(double sp) {
   double xosc = CC1101_CRYSTAL_FREQ * 1000;
 
-  int spMin = (xosc / (double)(1ULL << 18)) * (256. + 0.) * 1.;
-  int spMax = (xosc / (double)(1ULL << 18)) * (256. + 255.) * 8.;
+  uint32_t spMin = (xosc / (double)((uint32_t)1 << 18)) * (256. + 0.) * 1.;
+  uint32_t spMax = (xosc / (double)((uint32_t)1 << 18)) * (256. + 255.) * 8.;
 
   if (sp < spMin || sp > spMax) {
     return STATUS_INVALID_PARAM;
@@ -145,7 +145,7 @@ Status Radio::setChannelSpacing(double sp) {
 
   for (uint8_t e = 0; e <= 3; e++) {
     for (uint16_t m = 0; m <= 255; m++) {
-      double t = (xosc / (double)(1ULL << 18)) * (256. + m) * (double)(1ULL << e);
+      double t = (xosc / (double)((uint32_t)1 << 18)) * (256. + m) * (double)((uint32_t)1 << e);
       if (fabs(sp - t) < diff) {
         diff = fabs(sp - t);
         bestE = e;
@@ -180,8 +180,8 @@ Status Radio::setDataRate(double drate) {
   this->drate = drate;
 
   uint32_t xosc = CC1101_CRYSTAL_FREQ * 1000;
-  uint8_t e = log2((drate * (double)(1ULL << 20)) / xosc);
-  uint32_t m = round(drate * ((double)(1ULL << (28 - e)) / xosc) - 256.);
+  uint8_t e = log2((drate * (double)((uint32_t)1 << 20)) / xosc);
+  uint32_t m = round(drate * ((double)((uint32_t)1 << (28 - e)) / xosc) - 256.);
 
   if (m == 256) {
     m = 0;
@@ -208,8 +208,8 @@ Status Radio::setRxBandwidth(double bw) {
 
   */
 
-  int bwMin = (CC1101_CRYSTAL_FREQ * 1000) / (8 * (4 + 3) * (1 << 3));
-  int bwMax = (CC1101_CRYSTAL_FREQ * 1000) / (8 * (4 + 0) * (1 << 0));
+  uint32_t bwMin = (CC1101_CRYSTAL_FREQ * 1000) / (8 * (4 + 3) * (1 << 3));
+  uint32_t bwMax = (CC1101_CRYSTAL_FREQ * 1000) / (8 * (4 + 0) * (1 << 0));
 
   if (bw < bwMin || bw > bwMax) {
     return STATUS_INVALID_PARAM;
@@ -740,4 +740,3 @@ void Radio::waitReady() {
   while (digitalRead(MISO))
     ;
 }
-
